+++ KAIN2/Game/DEBUG.C
@@ -2873,7 +2873,7 @@ void DEBUG_FatalError(const char* fmt, ...)
 	while (1)
 	{
 		FONT_Flush();
-		DrawOTag((uintptr_t*)gameTrackerX.drawOT + 3071 * 2);
+		DrawOTag((unsigned int*)gameTrackerX.drawOT + 3071 * 2);
 		DrawSync(0);
 	}
 #endif
